requirements.txt
openai
sentence-transformers
pandas
qdrant-client
numpy
tqdm
python-dotenv
pyvi
